# affiliate/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Paparazzi Affiliate Commerce Platform — v1 API
  version: 1.0.0
  description: |
    Public v1 API for the Paparazzi Affiliate Commerce Platform (looks, offers,
    links, integrations, earnings, payouts, programmes, publishers, disputes,
    contracts, suspense).

    This spec is generated from the real route handlers in
    `packages/api/src/routes/*.ts` — it documents only endpoints the code
    registers. There is no `GET /v1/links` (links are minted, then resolved
    by the redirect service at `GET /r/{token}`).

    ## Authentication (temporary stub)

    All `/v1/*` endpoints require a Bearer JWT (`Authorization: Bearer <token>`).
    Expected claims: `sub`, `org_id`, `role`. **The JWT verification in
    `src/middleware.ts` (`requireAuth`) is a temporary stub: it trusts the
    token's claims verbatim and must be replaced by a production identity
    provider** with membership validation (`memberships` table), issuer /
    audience / expiry policy. `GET /healthz` is the only unauthenticated route.

    ## Envelopes

    - Success: `{ "data": <payload>, "request_id": "<uuid>" }` (`ok()` helper).
    - Error: `{ "error": { "code": "<ERROR_CODE>", "message": "..." }, "request_id": "<uuid>" }`.
    - Every response carries an `X-Request-Id` header echoing `request_id`.
    - Error `code` is always one of the `ErrorCode` enum values
      (`packages/shared/src/errors.ts`); clients should branch on `code`,
      never on message text.

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        TEMPORARY STUB — `requireAuth` in `src/middleware.ts` verifies the JWT
        signature against `JWT_SECRET` and trusts the `sub`, `org_id`, `role`
        claims verbatim. Replace with a production IdP (membership validation,
        issuer/audience/expiry policy) before pilot.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: |
        Optional idempotency key. First response per (key, org) is stored and
        replayed with `X-Idempotent-Replay: true`; 5xx is never stored.
...
  headers:
    X-Request-Id:
      description: Echoes the envelope `request_id` for log correlation.
      schema:
        type: string
        format: uuid
    X-Idempotent-Replay:
      description: Present (`true`) when the body is a replay of a stored idempotent response.
      schema:
        type: string
        enum: ['true']
  schemas:
    ErrorCode:
      type: string
      enum:
        - PROGRAMME_NOT_APPROVED
        - OFFER_STALE
        - PROPERTY_FORBIDDEN
        - PUBLISHER_NOT_ACTIVE
        - PROPERTY_NOT_OWNER_OPERATED
        - NOT_FOUND
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - CONFLICT
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - LEDGER_IMBALANCE
        - TRANSFER_STATUS_UNKNOWN
        - INTERNAL
      description: Canonical codes from `packages/shared/src/errors.ts`.
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
        request_id:
          type: string
          format: uuid
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data: {}
        request_id:
          type: string
          format: uuid
      description: Shape of every 2xx body; `data` is operation-specific.
    Uuid:
      type: string
      format: uuid
    MinorUnits:
      type: integer
      minimum: 0
      description: Integer minor units (paise for INR). Never decimals.
    Currency:
      type: string
      minLength: 3
      maxLength: 3
      example: INR
      description: ISO 4217 code; stored uppercase.
    DateTime:
      type: string
      format: date-time
...
    InternalConversionStatus:
      type: string
      enum: [received, pending, approved, declined]
    LedgerDisposition:
      type: string
      enum: [posted, skipped]
      description: |
        `posted` — ledger entries were written; `skipped` — conversion left
        unposted (unattributed click or no approved contract; never guessed).
    # ---- domain models ----
...
    ConversionEvent:
      type: object
      required: [provider_account_id, source_transaction_id]
      properties:
        kind:
          type: string
          enum: [conversion, reversal]
          default: conversion
        provider_account_id: { type: string, minLength: 1, maxLength: 200 }
        source_transaction_id: { type: string, minLength: 1, maxLength: 200 }
        line_id: { type: string, maxLength: 200 }
        returned_click_ref: { type: string, maxLength: 200, description: Matched against clicks.click_id; unknown stays suspense, never guessed. }
        currency: { $ref: '#/components/schemas/Currency' }
        eligible_value_minor: { $ref: '#/components/schemas/MinorUnits' }
        commission_minor: { $ref: '#/components/schemas/MinorUnits' }
        provider_status: { type: string, maxLength: 50, example: approved }
        occurred_at: { $ref: '#/components/schemas/DateTime' }
        provider_revision: { type: integer, minimum: 0, default: 0 }
        reversal_commission_minor:
          type: integer
          exclusiveMinimum: 0
          description: Required when kind=reversal.
        reason: { type: string, maxLength: 500 }
      description: |
        kind=conversion requires currency, eligible_value_minor,
        commission_minor, provider_status, occurred_at. kind=reversal requires
        reversal_commission_minor. Dedupe on
        (provider_account_id, source_transaction_id, line_id); a repeat at the
        same or lower provider_revision returns 200 `{deduped: true}`.
...
  responses:
    BadRequest:
      description: '400 — request failed validation (VALIDATION_ERROR).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: '401 — missing/invalid Bearer token or claims (UNAUTHORIZED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: '403 — role not allowed (FORBIDDEN), or a link-guard rejection (PROPERTY_FORBIDDEN, PROGRAMME_NOT_APPROVED, PUBLISHER_NOT_ACTIVE, PROPERTY_NOT_OWNER_OPERATED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: '404 — no route, unknown connector, or resource not in this org (NOT_FOUND).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: '409 — invalid state transition or over-reversal (CONFLICT), or imbalanced books (LEDGER_IMBALANCE).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: '500 — unexpected failure (INTERNAL); message not leaked.'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /v1/integrations/{connector}/events:
    post:
      tags: [Integrations]
      summary: Provider conversion/reversal webhook
      description: |
        Provider conversion notifications for the stub-network connector
        (`connector` path param allow-list: `stub-network`; anything else is
        `404 NOT_FOUND`). A `provider_account_id` of an Amazon.in Associates
        account, a reversal of an Amazon conversion, or a click ref of an
        Amazon link is `422 VALIDATION_ERROR` (Amazon rows come only through
        `POST /v1/integrations/amazon-associates/reports`).

        **kind=conversion** (default) runs the shared ingestion state machine
        (`src/conversion-ingest.ts`) — the same money path the CSV connector uses:
        - Attribution: `returned_click_ref` is matched against `clicks.click_id`.
          A miss is NOT an error — the conversion is stored with `click_id = NULL`
          (suspense). Attribution is never guessed.
        - Dedupe: `(provider_account_id, source_transaction_id, line_id)` is
          unique (nulls-not-distinct). A repeat at the same or lower
          `provider_revision` returns `200 { "deduped": true }`.
        - Revision ordering: higher revisions update provider_status/revision;
          `received/pending/declined → approved` posts the ledger (idempotent
          entry keys) and snapshots `contract_version_id`. An approval is NEVER
          downgraded — a higher-revision `pending` is logged and ignored, a
          higher-revision `declined`/`reversed` creates a reversal adjustment
          for the full remaining commission.
        - Approved conversions that cannot be attributed or contracted are
          logged and left unposted rather than posted against a guessed publisher.

        **kind=reversal**: partial/full reversal of a recorded conversion.
        `404` when the original is unknown; `409 CONFLICT` when
        `reversal_commission_minor` exceeds the unreversed remainder. Posts
        mirror entries snapshotting the ORIGINAL contract split.

        Roles: `network_admin`, `editor` (machine-to-machine).
      parameters:
        - name: connector
          in: path
          required: true
          schema: { type: string, enum: [stub-network] }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConversionEvent' }
            examples:
              conversion:
                value:
                  kind: conversion
                  provider_account_id: stub-merchant-001
                  source_transaction_id: txn-2026-09-20-0001
                  returned_click_ref: clk_9f2c1ab4
                  currency: INR
                  eligible_value_minor: 249900
                  commission_minor: 16000
                  provider_status: approved
                  occurred_at: 2026-09-20T10:15:30.000Z
                  provider_revision: 1
              reversal:
                value:
                  kind: reversal
                  provider_account_id: stub-merchant-001
                  source_transaction_id: txn-2026-09-20-0001
                  reversal_commission_minor: 8000
                  provider_revision: 2
                  reason: customer return within window
      responses:
        '202':
          description: Conversion created (or reversal recorded).
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        required: [conversion_id]
                        properties:
                          conversion_id: { $ref: '#/components/schemas/Uuid' }
                          adjustment_id: { $ref: '#/components/schemas/Uuid', description: Present on reversal / auto_reversed. }
                          status: { $ref: '#/components/schemas/InternalConversionStatus' }
                          auto_reversed: { type: boolean }
                          deduped: { type: boolean, example: true, description: Present on the 200 dedupe path. }
                          ledger: { $ref: '#/components/schemas/LedgerDisposition' }
              examples:
                created:
                  value:
                    data: { conversion_id: 66666666-6666-4666-8666-666666666666, status: approved, ledger: posted }
                    request_id: 9f2c1ab4-e5d6-4f7a-8b9c-0d1e2f3a4b5c
        '200':
          description: Deduped, status-changed, or auto-reversed (see `data` shape above).
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: |
            `VALIDATION_ERROR` — e.g. the conversion cannot be resolved to a
            programme (no attributable click and no single active programme for
            the connector), or a required field for the kind is missing.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/integrations/stub-network/payout-callback:
    post:
      tags: [Integrations]
      summary: Provider payout transfer callback
      description: |
        Stub-network payout rail callback: records the provider's outcome for
        one transfer (`provider_ref`) and rolls the batch status up:
        all `paid` → batch `paid` + ledger entries posted;
        any `failed` → batch `failed`.
        Roles: `network_admin`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_ref, outcome]
              properties:
                provider_ref: { type: string, minLength: 1, maxLength: 200, example: stub-9f2c1ab4-… }
                outcome: { type: string, enum: [paid, failed, unknown] }
      responses:
        '200':
          description: Callback applied; batch rolled up.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        required: [provider_ref, outcome, payout_batch_id, batch_status, ledger]
                        properties:
                          provider_ref: { type: string }
                          outcome: { type: string, enum: [paid, failed, unknown] }
                          payout_batch_id: { $ref: '#/components/schemas/Uuid' }
                          batch_status: { type: string, enum: [processing, paid, failed] }
                          ledger: { type: string, enum: [posted, skipped, n/a] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...